# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Instagram-clone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories:
    post:
      operationId: createStory
      tags: [stories]
      summary: Create a new story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Story created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{storyId}:
    delete:
      operationId: deleteStory
      tags: [stories]
      summary: Delete a story (author only)
      parameters:
        - name: storyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /stories/{storyId}/view:
    post:
      operationId: viewStory
      tags: [stories]
      summary: Mark a story as viewed
      parameters:
        - name: storyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Viewed
  /stories/{storyId}/viewers:
    get:
      operationId: getStoryViewers
      tags: [stories]
      summary: Get viewers of a story
      parameters:
        - name: storyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Viewers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /stories/{storyId}/react:
    post:
      operationId: reactToStory
      tags: [stories]
      summary: React to a story with an emoji
      parameters:
        - name: storyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryReactionInput"
      responses:
        "200":
          description: Reaction recorded
  /stories/upload:
    post:
      operationId: uploadStoryMedia
      tags: [stories]
      summary: Upload media for a story (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadInput"
      responses:
        "200":
          description: Uploaded media URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadMediaResponse"

  # ── Story Highlights ──────────────────────────────────────────────────────
  /stories/highlights:
    post:
      operationId: createHighlight
      tags: [stories]
      summary: Create a story highlight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                coverUrl:
                  type: string
                storyIds:
                  type: array
                  items:
                    type: string
              required: [name]
      responses:
        "201":
          description: Highlight created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
  /stories/highlights/user/{username}:
    get:
      operationId: getUserHighlights
      tags: [stories]
      summary: Get highlights for a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Highlights list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Highlight"
  /stories/highlights/{id}:
    delete:
      operationId: deleteHighlight
      tags: [stories]
      summary: Delete a highlight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /stories/highlights/{id}/stories:
    post:
      operationId: addStoryToHighlight
      tags: [stories]
      summary: Add a story to a highlight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storyId:
                  type: string
              required: [storyId]
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]

  # ── Notifications ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isFollowing:
          type: boolean
      required:
        - id
        - username
        - fullName
        - isFollowing
...
    Story:
      type: object
      properties:
        id:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        audience:
          type: string
          enum: [everyone, close_friends]
        viewsCount:
          type: integer
        isViewed:
          type: boolean
        author:
          $ref: "#/components/schemas/UserSummary"
        expiresAt:
          type: string
        createdAt:
          type: string
        caption:
          type: string
          nullable: true
        textColor:
          type: string
          nullable: true
        musicUrl:
          type: string
          nullable: true
        musicTrack:
          type: string
          nullable: true
        musicArtist:
          type: string
          nullable: true
      required:
        - id
        - mediaUrl
        - mediaType
        - audience
        - viewsCount
        - isViewed
        - author
        - expiresAt
        - createdAt
...
    StoryInput:
      type: object
      properties:
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        audience:
          type: string
          enum: [everyone, close_friends]
        caption:
          type: string
          nullable: true
        textColor:
          type: string
          nullable: true
        musicUrl:
          type: string
          nullable: true
        musicTrack:
          type: string
          nullable: true
        musicArtist:
          type: string
          nullable: true
      required:
        - mediaUrl
        - mediaType
    StoryReactionInput:
      type: object
      properties:
        emoji:
          type: string
      required:
        - emoji
    Highlight:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        coverUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        stories:
          type: array
          items:
            $ref: "#/components/schemas/Story"
      required:
        - id
        - userId
        - name
        - createdAt
        - stories

    # ── Notifications ────────────────────────────────────────────────────────
...
    MediaUploadInput:
      type: object
      properties:
        data:
          type: string
          description: Base64 encoded file data
        mimeType:
          type: string
      required:
        - data
        - mimeType
    UploadMediaResponse:
      type: object
      properties:
        url:
          type: string
        publicId:
          type: string
        resourceType:
          type: string
      required:
        - url
        - publicId
        - resourceType
...